Make CompFlow/Problem::initialize() generic
This commit moves up the node-specific assignment of the initial
conditions or (the analytic solution if exist for a problem) from the
specific CompFlow/Problem policiy classes to cg::CompFlow::initialize().
This enables writing the equivalent elem-specific assignment of the ICs
(and the analytic solution if exist) in dg::CompFlow::initialize(). From
these functions we still call down to the problem-specific setting of
the ICs (and the analytic solution if exist) via Problem::solution().
This is accomplished by exposing this latter member function of all
specific CompFlow Problems as public and with the same signature, so
they can be called in a generic fashion.