Make CompFlow/Problem::dirbc() generic
This commit moves up the node-specific query for Dirichlet boundary
conditions (BCs) from the specific CompFlow/Problem policiy classes to
cg::CompFlow::dirbc(). This enables writing the equivalent elem-specific
setting of BCs in dg::CompFlow::dirbc(). From these functions we still
call down to the problem-specific evaluation of the solution increments
(for those problems where an analytical solution exist). This is
accomplished by exposing this latter member function of all specific
CompFlow Problems as public and with the same signature, so they can be
called in a generic fashion as Problem::solinc().

This commit renders the CompFlow/Problem policy classes ready for both
CG and DG.

The last step towards enabling the PDE machinery to work with DG is to
teach the parser's grammar to configure the right PDEs based on the
already existing parsing for the user input for the discretization type.