Send only chare-boundary surface nodes/edges during inciter setup
This commit reduces the volume of communication from domain to surface
nodes and thus leads to a drastic improvement of the setup phase of
inciter, especially with mesh refinement.

In particular, this commit puts in a preparatory step in
Partitioner::gather(), which marks the beginning of the setup phase
before global distributed mesh reordering can start.

This setup phase consists of a handshake, a two-step procedure, in which
all PEs send a query to all other PEs with the node IDs of the mesh
chunk they hold after partitioning. In the second step of this
handshake, as an answer to the query, each PE responds to all other PEs
with the node IDs and the associated chare IDs they share with the PE
the request came from. The two steps of this handshake are given by
Partitioner::query() and mask(), respectively.

Before this commit the node IDs (and edge-end-point IDs in case of
initial uniform mesh refinement was selected by the user) were sent
during the query phase for the full mesh chunk domain. This commit
replaces the domain node IDs with only those node IDs that are on the
chare-boundary surface, which are the only ones that are shared among
chares and PEs. Thus this reduces the amount of data transferred during
this two-step all-to-all communication step.

Tested with a 18M-cell mesh this yields about a 10% improvement without
mesh refinement (which does not send edge-end-point IDs), but leads to a
serious speedup of 3.6x with mesh refinement. This is because the
edge-end-point IDs contain a lot more data compared to node IDs.

Other interesting details:

  * For testing purposes, temporarily switched back to the old
Partitioner::refine(), which is faster that the newer and more general
one that is under construction. Note that the more general refine() is
still there but commented.

  * Clarified the documentation on and abstracted away the ordering
convention in lpofa, which is now reused in multiple places.