Read mesh node coordinates one by one in inciter
Somewhat counterintuitively, this does not make reading the coordinates
slower but it allows inciter using a lot less memory when reading the
mesh node coordinates in parallel.

Since in a parallel run reading the mesh is distributed by elements,
i.e., PE0 reads approximately elements 0...n, PE1 reads approximately
elements n+1...2n, etc., node coordinates of all node IDs associated to
elements that are read must be read in.  However, there is no reason to
assume that the node IDs of the elements read are nicely contiguously
numbered (and experience shows that they are usually not). Thus so far
inciter has computed the extents, i.e., the minimum and maximum, of the
global node IDs and read in all nodes between them, in the hope that
reading in multiple node coordinates will always be faster than reading
them one-by-one. This however, is not very good since there is still a
lot of nodes that get read in such a way, because the nodes are so
arbitrarily badly ordered in the file compared to the consecutively read
elements. As a result, most PEs read most of the coordinates, which does
not scale. Thus this commit replaces this extent-based node reading
strategy with a more memory-efficient one, which reads only those nodes
that are absolutely necessary on a given PE, reading node coordinates
one-by-one but with global node IDs in increasing order, which other
than that though, are arbitrary.

The symptom of this problem came when a 144M-cell mesh was fine on 25
compute nodes (each with 128GB RAM), but the nodes of a 288M-cell mesh
could not be read on 25 nodes and not even on 70. As a result of this
commit, a 288M-cell mesh can now be read fine on 25 nodes (with or
without initial uniform mesh refinement, which of course, uses
additional memory).