Handle errors with backtrace using signals
This allows generating nice demangled (at least on linux) backtraces not only
from exceptions, but also from places throwing exceptions that are in Charm++
chares, from which catching exceptions cannot be done in an obvious C++ fashion,
I guess, due to the asynchronous Charm++ machinery.

The hope is that these signal handlers, now set for all executables as early as
we can, will always generate (1) at least a line number where the problem
occurred, and (2) also hopefully a readable backtrace.