Turn off backtrace on Alpine Linux
When exceptions are thrown or when a signal is caught, we want a
backtrace so we can diagnose problems.

For some reason I cannot debug further, execinfo's backtrace() function
call segfaults on alpine linux. This commit disables backtrace by
detecting the OS and turning backtrace off on alpine.

It appears that on alpine backtrace never worked for us, but we also
never tested it. ;-) Previously, it was disabled on RELEASE builds
(which we test) and now that I want to enable backtraces independent of
DEBUG vs RELEASE (or whatever), this breaks, so this commit puts in an
alpine-specific fix.

I'm not very happy with this solution. For example, detecing execinfo
instead would have been better, but alpine's quick'n'dirty
implementation of gnu's libexecinfo package does not provide a compiler
macro for detecting execinfo, so we disable backtrace on alpine.

References:

http://man7.org/linux/man-pages/man3/backtrace.3.html
https://pkgs.alpinelinux.org/package/edge/main/x86/libexecinfo-dev
https://www.freshports.org/devel/libexecinfo
https://www.gnu.org/software/libc/manual/html_node/Backtraces.html
https://github.com/unbit/uwsgi/issues/1514 (may be related)