Use policy classes for injecting state into DG boundary surface integrals
This eliminates the if-test for the portion of computing the boundary surface
integrals that is specific to setting the left and right hand state of the
surface but still allows code reuse.

Also update GaussHump regression test excercising DG advection to incorporate
symmetry, inlet, and outlet boundary conditions on arbitrary number of side
sets.

Also remove DGPDE::dirbc(), as this appears to be unnecessary now.