Search among boundary instead of domain nodes in Partitioner::mask()
This commit replaces the search among domain nodes with a search among
only the chare boundary nodes in Partitioner::mask(). Accordingly,
Partitioner::m_nodechares is now renamed to m_bnodechares. This data
structure is a map that associates chare IDs to (potentially multiple)
boundary node IDs on each PE and is used during global distributed mesh
node reordering in Partitioner to setup two communication maps, m_msum
and m_ncommunication. The former is a chare-node communication map and
it is symmetric, while latter is a PE-node communication map and is
asymmetric. The chare-node comm map is used later during solve, the
PE-node comm map is used during node reordering only.

Another subtle but important change is to move the computation of the
chare boundary nodes (in bnodechares) from gather() to flatten(). Since
in mask() now we search in this data structure, this is necessary to
ensure that this data is ready before mask() is called. Since flatten()
is followed by a global sync, this ensures that all PEs have a complete
bnodechares data before gather(), query, and mask() can be called.

There is only a minor performance improvement due to this change
measured using a 14MM-cell mesh on 288 PEs.

While this commit does not break the AMR tests, it does eliminate some
code that will not be needed in anticipation of introducing the AMR
object into Partitioner before Zoltan. So this is also a preliminary
cleanup step.