Add inciter::MeshReader
For now MeshReader can read from ExodusII mesh files but this is where
Omega_h's osh format can also be added in the future abstracting this
from Partitioner.

The main addition of this commit is to revamp Partitioner's constructor
with #220 in mind. In particular, we now just read in the mesh node
coordinates right after the mesh graph (connectivity). This simplifies
the code as we don't expect to gain much from delaying the node
coordinate read.

This commit also puts in some test code under Partitioner::refine()
which successfully uses the AMR object to (1) do a single-level uniform
refinement before Zoltan, (2) extract the new connectivity, (3) compute
the new coordinates for the newly added nodes (assuming single level
refinement), (4) and save the new mesh for verification. This will, of
course, change based on our progress according to the plan roughly
outlined in #220.

Also add new constructor to mesh_adapter (this is all, sorry for the
white spaces).

Partitioner now stores the node coordinates in its state, which will be
updated after reordering.

The changes in Transporter are due manly to slightly reorganizing the
screen output during the mesh prep step, consisting of (1) the mesh read
(both graph and node coords), (2) optional refinement, and (3) computing
the centroids (if needed). Accordingly, the DAG wait wait4part is now
renamed to wait4mesh, which better expresses that we are preparing the
mesh waiting for that label to be true. When that is done, we proceed to
mesh partitioning, calling Transporter::partition(). Also, a better
place to output mesh graph statistics and load distribution on the
screen is in Transporter::partition(), which only happens after the
initial refinement at which we know exactly how many new mesh cells have
been added during a potentially non-uniform initial refinement step.

Note that this commit breaks the amr regression tests (but supposed to
break only those) as we are WIP towards AMR before Zoltan which will
eventually even require updating the baselines for these tests.