Add hard-coded multilevel initial refinement
This commit adds a four-level initial mesh refinement if the user
requested initial AMR. This is still serial only. The number of levels
are hard-coded at this point, the goal here was to just see how
multi-level uniform refinement can be accomplisehd through the AMR
object.

Also replaced mesh_adapter_t::refine with a reference and removed the
pesky single call to operator new from AMR.