Convert intiial amr types to a list, add uniform levels
This now allows things like:

  amr
    initial uniform
    initial uniform
    initial ic
    initial uniform
    initial ic
    initial ic
    initial uniform

    uniform_levels 2
    #error hessian
  end

Which (1) sets the levels if initial uniform refinement to 2, and (2)
grows a list of "initial refinement types" that will be performed in
that order before t=0. Still need to weed out duplicate 'ic' entries as
that makes no sense.

Also temporarily disable amr regression tests.