Require depvar for both transport and compflow
'depvar' throughout the code is used to identify a system of equations among
multiple systems. This is both used in walker for DiffEq systems as well as in
inciter for PDE systems. Each system has a unique single-character identifier,
depvar, which can also be thought of as the dependent variable in the
mathematical sense. The depvar is given by the user in the input file and now
must be defined for all systems. In walker, for example, the depvar is used to
specify statistics. In inciter, for example, it will be used to identify scalar
field(s) designated as variable that trigger mesh refinement.

This commit also augments some of the regression tests for inciter that did not
specify depvar for either Transport of Compflow systems of PDEs. Previously,
depvar was not mandatory, now it is an error if unspecified.