Parse refvar ... end block for initial amr
These refvars will be used for estimating the numerical errors in the initial
conditions specified for the scalar components which then will be used to
trigger initial mesh refinement.

This commit puts teaches inciter's parser to recognize a new refvar ... end
block within inciter's amr ... end block. Refvar stands for refinement
variables, the block parses an arbitrary list of dependent variables (depvar).
Depvars consist of a character that identifies a system among multiple system of
equations optionally followed by a field ID. Example: c, d2, which denote,
respectively, the first scalar component of the equation system whose depvar is
denoted by c and the second component of another equation system whose depvar is
d.

Depvars must be unique, must be defined upstrea of the refvar ... end block
where they are referenced, and their field id must be between 1 and the number
of scalar components configured for the given system. All this is ensured by the
parser, in case any of this is not satisfied, parser errors are generated.

The meat of this commit is in the functor compute_refvar_idx, which computes
absolute indices (locations) of the refvars specified for the system of systems
configured upstream of the refvar ... end block. The result of the functor is
stored in the index vector under inciter::g_inputdeck.get<tag::amr,tag::id>().
This, however, required fixing the offsetmap generated in ncomponents, which was
previously wrong. Note that this offset map is also, and so far has only been,
used by walker when computing statistics for indexing into the vector of
unknowns for a given depvar, which has been wrong for systems, however, neither
the regression tests for walker nor how walker has been used so far involved
systems, so this has not come out. This is now fixed, together with its unit
test. Another piece for computed the refvar indices of initial amr was the
generation of another map that associates the the number of components in a
system to its depvar. This is so that we can do proper error checking on the
field id specified during parsing.