Fix call to Transporter::comfinal
There was a bug in Solver::created(), which this commit fixes. When
chares called this function (signaling that communication maps have been
completed) it immediately forwarded the call to Transporter::comfinal(),
which is wrong for nonzero virtualization. The correct way is to wait
for all chares contributing to created() on a given PE and only then
fire up the reduction to Transporter::comfinal(). This ensures that all
chares really have finished computing their communication maps.

Note that this affected DiagCG as well as DG, but DiagCG needs no
communication maps (other than what's already available in its base,
Discretization), so DiagCG simply calls created() in its ctor. However,
DG does need to compute extra comm maps, and that's what tripped on this
error as some of the DG chares were not ready in parallel + nonzero
virtualization.

Another change is the removal of DG::ready4ghost, after realizing that
calling ready4ghost() was really doing the same logic as reqGhost().
Thus the former is superfluous and latter can do this job if (and only
if) it is called after after setupGhost(), which is how it is.

We also do not call created() from DG, but it appears to be enough to
start the contribute to Transporter::comfinal from Discretization(),
hence d->contribute(). (I really wish DiagCG could also just do this,
but that cannot.) Note that this now eliminates the need to store
m_solver in DG, which is nice, because DG does not need a solver. Also
note that taking the d-> out in front of contribute() in DG will cause
some of the DG regression tests to dead-lock. This has to do with
Charm++ not accepting multiple reductions from the same chare array
fired up in arbitrary order. If that may happen, one can start some of
the reductions from a shadow array, e.g., Discretization, which is what
we do here buy doing d->. (Or as it has been so far, call to Solver,
which is a group, and originate the reduction from there, as is done
from DiagCG.)

Other minor cleanup.