Correctly wait for both own and communicated part of the unknowns before DG::solve
Previously, it was incorrectly possible that the owned contribution to the
unknown/solution vector was not ready and the solution was done when all ghost
data has been received. This commit ensures that both own and communicated,
i.e., ghost, parts of the unknown/solution vector are complete on a chare before
the solution. As usual, this is done with a Charm++ SDAG entry in dg.ci.