Add missing MatCG::m_side to pup routine
With this change MatCG appears to work with migration but there is a couple of
problems: (1) Testing for correctness of Dirichlet BCs in MatCG::correctBC(),
incorrectly triggers sometimes, and (2) even if it does not, employing
RandCentLB eventually appears to scramble the chares significantly so that
communication required, done by Solver, before the linear solve gets
unacceptably slow.

(1) is probably a bug (turning off BCs makes it run) while (2) is long expected.
Both of these issues will need to be investigated before migration in MatCG's
will be useful.