WIP towards parallel initial mesh refinement
At this point we have a setup in Partitioner that is capable of doing a
series of multiple initial mesh refinement steps in parallel. In other
words, after the reading the mesh in a parllel distributed fashion, we
enter a loop in which we (1) do an initial mesh partitioning calling
Zoltan without virtualization, (2) distribute the mesh (cells) according
to the result of the partitioning, (3) output the mesh, and, (4)
consuming the list of initial mesh refinement steps the user has
configured, we continue by a new partitioning step to step (1). We exit
if the list of initial refinement steps hav ebeen consumed. No
refinement is enabled at this point, and the mesh gets repartitionied
and redistributed among the PEs (supposedly correctly, the mesh for now
correctly does not change after the first one).

Next is to enable refinement and make the PEs agree on a conforming mesh
across PE boundaries.

At this point we abort before calling Partitioner::flatten(). The plan
is to call Zoltan one more time after the list of initial mesh
refinement steps has finished, but now enabling overdecomposition.