Initial mesh refinement in parallel
Appears to be correct in parallel using multi-step initial uniform
refinements. Zoltan correctly repartitions the mesh in every step,
followed by a step that redistributes the mesh to their PEs assigned by
the repartitioning, We then generate the edges on PE boundaries and each
PE distributes them to those PEs they communicate with. Then the mesh is
refined, resulting in new edges. The newly added nodes on edges are
given a new global unique ID using a combination of the hashes of the
parent (global node) IDs of the edge. The edges on the PE boundaries are
then communicated to other PEs in order to verify that (1) the new
global node IDs, independently generated on different PEs, are the same,
and (2) their coordinates are also the same to machine precision, which
then means they are really the smae physical node on mutliple PEs that
share the refined edge. Note that the coordinates will most likely not
be needed in the future, but they are there for now for verifying
correctness.

The next step is to make sure the above also works with not just a
series of uniform refinement steps, but also with multiple
error/IC-based nonuniform refinement steps. That will result in edges on
PE boundaries where a PE added a new node but another one did not, for
which we still need to connect the mesh in a conforming fashion.

Eventually, all this initial mesh refinement functionality will be moved
out from Partitioner to a separate class.