Connect the result of initial mesh refinement
This commit connects the result of the initial mesh refinement back to
the code so that operators can compute PDEs on it.

The main change is the communication of the mesh node coordinates after
the partitioning/reshuffling step after the multiple steps of initial
mesh refinement. So far this step only dealt with node IDs, but now we
also do the coordinates. Appears to work, but not complete yet:

1. Non-uniform refinement produces non-positive Jacobians even in
serial.

2. Works in parallel, but not yet with overdecomposition:
Discretization::setCoord() triggers an assert.

Beside fixing the above issues, it is now also time to re-enable the
initial-AMR regression tests.