Skip initial mesh refinement if not typelist defined
This change enables skipping the initial mesh refinement step if the
initial mesh refinement type list is empty. The amr...end block can
still be defined, but if the type list is empty we now correctly skip it
and proceed to the chare distribution step.

Partitioner::m_gelemid has also been removed from Partitioner's state as
it is only needed locally for Zoltan.