Ensure multiple chares can contribute to a node
This was still required for computing the chare-categorized mesh node
coordinates. This now ensures that all the correct node coordinates are
transferred to Discretization worker chares and thus is correct with
nonzero virtualization. (Previously, not all nodes shared by multiple
chares went to chares that shared nodes, thus the assert, now this is
fixed.)

Now the regression tests run and pass with optional initial mesh
refinement in place. Next is to re-enable the regression tests that
exercised initial AMR, which will now change even for a single step of
uniform refinement, as we now do the mesh refinement before mesh
partitioning, to ensure better initial static partitions.