Move testing for the existence of cells on a PE to after reshuffling
Previously, an assert was triggered if there were no cells assigned to a
PE by the partitioner on a given PE. This is not necessarily a problem.
The intention was to detect PEs which had no mesh cells assigned to them
at all, which can be detected after reshuffling, which is what this
change does.