Evaluate IC for initial mesh refinement for CG and DG
Previously, initial mesh refinement triggered based on the numerical
error estimated in edges was only hooked up for nodal discretizations,
e.g., MatCG or DiagCG. This commit puts in code that detects the
discretization type (node or cell-centered), evaluates the ICs correctly
in nodes or cells, and if the solution is given in cells, it transfers
the solution to nodes, by simply averaging the fields. This nodal field
is then used, as before, to estimate the numerical error in the initial
conditions, and tag edges fro IC/error-based initial mesh refinement.

No regression test with DG + IC-based initial mesh refinement, because
the latter produces negative Jacobians at this point.