Transfer cell values to nodes weighing by volume
This is for IC/error-based initial mesh refinement for cell-centered
discretizations, such as DG. Previously, we trasnferred the IC field
values from cells to nodes by a simple arithmetic average from cells
surrounding a node (and forgot to divide by the number of cells).  Now
this us a sum weighed by the cell volume, which is more suitable for
variables conserved in a control volume.

Note that this is incorrect in parallel at PE boundaries since the sum
does not consider cells surrounding a node across PEs. However, this
only affects refinement criteria, i.e., wether an edge is tagged for
refinement or not. Transferring data from cells to nodes correctly
requires communication and we will see we can get away without it (only
for the purpose of tagging edges for refinement) for now.