Put in code to update boundary data structures after initial mesh refinement
This addresses the last two bullets of #236, and 861f9b8f hopefully
addresses the first bullet. So what's left for initial mesh refinement
to be complete is the second bullet in #236: cleanup mesh refinement
stuff away from Partitioner to its own class which then will be also be
reused for t>0 refinement.

This change was only verified with a coupled of simple examples, but
exercising multiple levels of initial refinement, involving both uniform
and ic-based refinement, in serial and parallel, and using two different
meshes: appears to correctly work all around.

Updating the boundary face data structures enables both CG or DG
discretizations to set boundary conditions correctly on the refined
mesh.  More testing and new regression tests are forthcoming.