Walker: Add Gaussian initpolicy
This enables all types of SDEs walker supports to be initialized using a
joint Gaussian composed of independent marginal Gaussians, i.e., no
correlations.  Example usage in the walker...end input block:

  langevin
    depvar u

    init jointgaussian
    icgaussian
      gaussian 0.0 5.0 end
      gaussian 1.0 1.0 end
      gaussian -2.0 0.5 end
    end

    C0 2.3
    coeff hydrotimescale
    hydrotimescales eq_A005S end
    hydroproductions prod_A005S end
    rng rngsse_gm55
  end

Which configures a Langevin equation, whose

* dependent variable is 'u',

* has three scalar components (implicit) as intended for velocity,

* selects the joint Gaussian initialization policy,

* configures three Gaussians for the three components with means 0.0,
1.0, -2.0 and variances (=sigma^2) 5.0, 1.0, and 0.5, respectively,

* setting C0 in in the Langevin model as 2.3 (overriding the default
2.1),

* configures the coefficients policy to be 'hydrotimescale', which pulls
in DNS data for k/eps from homogeneous Rayleigh-Taylor for nine
different configurations,

* configures the DNS timescales and prod/eps for the symmetric Atwood
number 0.05 case,

* and selects the RNGSSE GM55 random number generator.

The new grammar that is being added to walker is to recognize the
Gaussian-related input above. Parser error checking is done only on
accepting exactly two integers between the gaussian...end block, no
attempt is made to match the number of configured components for a given
SDE with the number of configured marginal Gaussians.

Also improved error message for when a DiffEq type is being instantiated
but not found.

Next is to setup a case and add a regression test for using this as the
simplified Langevin model for a simple homogeneous decay, and verify the
solution.