Add coupled position and velocity model
This adds a deterministic Lagrangian position model that can be coupled
to the Langevin model for the velocity in walker.  Tested with
homogeneous shear, appears correct, except the single nonzero
off-diagonal component of the Reynolds stress anisotropy for this flow,
which needs further debugging.

Some error checking is done for parsing of the position...end and
langevin...end block, configuring the position and velocity model,
respectively, and their coupling.

There is also a workaround put in for when querying the position of the
PEGTL input at the end (or after the end?) of the input stream (at the
end of parsing the file), which segfaults, when attempting to add an
error message using that input location. In that case we don't generate
location info but still generate an error message but outside of the
normal parser error generation mechanism.

Note that this is highly experimental and still needs regression tests,
example setups, and documentation.

Example input file, testing homogeneous shear with the coupled position
and simplified Langevin model for the velocity:

=========

title "Coupled position and velocity problem computing a homogeneous shear flow"

walker

  #nstep 2     # Max number of time steps
  term  15.0    # Max time
  dt    0.1   # Time step size
  npar  100000 # Number of particles
  ttyi  1    # TTY output interval

  rngs
    rngsse_gm55 end
  end

  position      # configure a position equation
    depvar x
    velocity u  # couple a velocity model with dependent variable u
    init zero
    coeff instantaneous_velocity
  end

  langevin      # configure a Langevin velocity equation
    depvar u
    position x  # couple a position model with dependent variable x
    init jointgaussian
    icgaussian # unit kinetic energy, isotropic Reynolds stress at t=0
      gaussian 0.0 0.666667 end
      gaussian 0.0 0.666667 end
      gaussian 0.0 0.666667 end
    end
    C0 2.1
    # specify coefficients based on simplified Langevin model and
    # prescribed shear
    coeff const_shear
    rng rngsse_gm55
  end

  statistics
    interval 1
    # estimate and save mean velocity
    <U1> <U2> <U3>
    # estimate and save Reynolds stress
    <u1u1> <u2u2> <u3u3> <u1u2> <u1u3> <u2u3>
  end

  pdfs
    interval  1
    filetype  txt
    policy    multiple
    centering elem
    format    scientific
    precision 6
    # save marginal PDFs for all velocity components
    u1( U1 : 1.0e-2 )
    u2( U2 : 1.0e-2 )
    u3( U3 : 1.0e-2 )
  end
end

=========

The time evolution of the Reynolds stress anisotropy can be verified
with the following gnuplot command:

plot "stat.txt" u 2:($6/($6+$9+$11)-1.0/3.0) t "b_{11}", "" u 2:($7/($6+$9+$11)) t "b_{12}", "" u 2:($9/($6+$9+$11)-1.0/3.0) t "b_{22}", "" u 2:($11/($6+$9+$11)-1.0/3.0) t "b_{33}"