AMR chares (Refiner) now created and passed a mesh
This change basically puts in a fork for optional initial mesh
refinement after mesh partitioning and reshuffling and before global
node reordering. For now only the Refiner objects are dynamically
inserted and the initial mesh refinement step is shortcut.

An improvement is that now we create specific callabck tagged tuple
types in the newly added file Base/Callback.h. This allows extending
these types at a single location and produces a compile-time error if
they are not initialized correctly. Before this was more error-prone
because the callbacks were converted to a vector and their transfer to
inside Partitioner had to be maintained separately. Now that we have
another bundle of callbacks passed in and forwarded to refiner, the
previous approach would be harder to keep consistent at various places
when it needs a change.