Hook back up boundary edge distribution in AMR
Shortcut immediately after for now. Also make tk::Progress more
user-friendly by requiring to pass the list of legends of subtasks in
the constructor - this way the prefixes and the legends are at the same
place in client code and thus also easier to read and know what the
prefixes mean.

Temporarily disable all AMR regression tests.