Add Sorter, a class used for mesh reordering
This is a step towards revamping inciter's setup and mesh preparation
stage. After refiner optionally refined the mesh, it hands it over to
this new class, Sorter, which does global distributed mesh node
reordering. That node reordering at this point is skipped and Sorter
instead sets up a node-communication map and creates the workers (base
and child) as before.

All of this significantly changes the asynchronous logic, which appears
correct and perhaps better compared to what it was.

This is tested in serial and parallel, using different meshes, and
either skipping initial mesh refinement or doing a single uniform step.
Everything appears correct, and functional. This is mostly tested with
DiagCG, MatCG definitely does not work yet, as the linear system bounds
are not computed.

Next is to enable node reordering.

This commit would fail most of inciter's regression tests.