Make meshconv generate triangle block if only tet exist
This enables meshconv, using exo as both input and output, to detect if
a triangle element block does not exist and generate one. It also
transfer the side sets from the input exodus file.

There is a bunch of related changes:

1. Remove an inciter regression test that was put in to use a mesh that
does not have triangle element block. This is now required. For DG, it
is not okay to have no triangle element block or node side sets. For CG,
it is okay to not configure a side set for a problem. (Some inciter
regression tests use meshes without side sets, so the absence of side
sets and/or triangle block in the input file is not made an error, only
for DG.)

2. Modify a bunch of meshconv regression tests. These were necessary as
meshconv now adds a triangle element block if it does not find one and
the output file diffs if there is a new element block.

3. Changed the order of triangle and tet element blocks as
ExodusIIMeshWriter saves them. This was needed to transfer the side set
tet ids from the input and output mesh when adding a triangle element
block. As a result, more meshconv regression tests had to updated.
Visually verified the new gold standard meshes to be correct.

4. Fixed a couple of bugs in ExodusIIMeshReader:

  * m_nel was not cleared between potentially successive calls to
  readElemBlockIDs()

  * readFaces() was attempting to read as many triangle elements as many
  boundary faces are found assigned to side sets. These are not
  necessarily always the same. The correct number of triangle elements to
  read is the max number of triangle elements in the file.

5. ExodusIIMeshWriter now write side sets, but only if a tk::UnsMesh
object is passed to it. This is how meshconv uses the mesh writer, but
not how inciter uses it. If side sets are needed that takes a side sets
themselves, a new API can be added. This is consistent with how inciter
reads and writes meshes and field data into exodus files: it writes
different components independently as (1) that scales better in parallel
and (2) it is more convenient to read/write components when they are
needed/ready instead of creating a mesh object with everything inside.

6. Remove the piece of code that reads the node lists associated to side
sets if no triangle element block is found in the input file: this is in
two places: Partitioner::createWorkers() and
Transoprter::createPartitioner(). For DiagCG/MatCG if BCs are desired,
sidesets are required, which are obtained from triangle element
connectivity and side sets instead of attempting to read the node lists.
DG always requires side sets as well as triangle element connectivity.

7. Enabled some commented out exception catches, that were erroneously
left commented some time ago.

8. Since readSidesetFaces() slightly changes, some unit tests also had
to be updated.