Update node lists after initial mesh refinement
This commit puts in code to correctly regenerate node lists associated
to side sets after initial mesh refinement. Still only tested with a
single uniform refinement, but appears to work correctly. This is
without node reordering, which now will be possible to turn off - no
need to pay for it for DiagCG or DG.

To get the node lists for side sets, we now use the ExodusII API instead
of generating them from face data (i.e., outsourced). We do not even
read the face list and triangle connectivity if CG is configured.

This is only tested with DiagCG so far, MatCG, DG do not yet work as not
everything is hooked back up.

Also adds more verification code, e.g., test for leaky mesh partition
before and after mesh refinement, (re-)using the code that was
previously added to DG only, now moved to DerivedData,
tk::leakyPartition().