Hook back up face data after initial refinement
With this commit DG now also appears to work and compute the correct
result in both serial and parallel with and without initial mesh
refinement. (Exercised with a single step of uniform refinement for
now.) DiagCG also works, MatCG is next. This is still without mesh
reordering. DiagCG and DG regression tests pass.