Hook back up mesh node reordering
This commit hooks back up global distributed mesh node reordering
(PE-locality mesh node reordering) that yields consecutive unique global
node IDs with increasing PE IDs. This is important for MatCG, DiagCG and
DG do not care about it.

This commit also teaches inciter's parser to understand the following
input:

inciter

  reorder true  # or false

end

which turns this reordering on/off. The default is off, unless MatCG is
configured by the user in which case it is on and the reorder user
setting has no effect.

All of this reordering is now in the new class Sorter, and that's all
Sorter does: after (optional) mesh refinement, Sorter reorders the mesh
nodes, but only if configured (and always if MatCG is configured).

While the previous mesh refinement regression tests have still not been
re-enabled, all regression tests now pass again, even though this global
reordering is now NOT performed by default for DiagCG and DG.

The code that does the PE-locality mesh reordering (done in Sorter) is
significantly simplified compared to how it was in Partitioner. This is
mostly due to the fact that Sorter is a Charm++ chare array, and
Partitioner is a Charm++ chare group. Of course, Sorter can also be
migrated, thus even the PE-locality mesh reordering should be possible
to load-balance using Charm++ in the future.

Removed some of the documentation of SDAG, as it is pretty cumbersome to
maintain, and the Charm++ SDAG code is a lot easier to read and
interpret that reading through an essay.

Removed linear system communication cost estimate. This wasn't very
useful anyway and complicated the asynchronous logic quite a lot.

With this, the revamp of the setup of inciter is complete. This is
tested with only a single step of uniform initial refinement (on/off),
reordering (on/off), serial/parallel, zero/nonzero virtualization, and
all three of MatCG, DiagCG, and DG - all combinations of these are now
supposed to work.

Next steps:

* Enable non-uniform (IC-based) initial mesh refinement

* Enable multiple sets of initial mesh refinement

* Add regression tests covering all the above combinations with
  - serial, parallel, parallel + virtualization
  - reorder, no reorder
  - no ref, uniform ref, IC-ref, arbitrary multiple ref steps
  - DiagCG, MatCG, DG