Walker: add gamma initpolicy and gamma RNGs
This commit adds to all types of random number generators the ability to
generate from a gamma distribution, whose shape and scale parameters are
configured by the user. This then can be used as an initialization
policy for any of the equations in walker.

The generated distribution is visually verified via the dissipation
equation configured using the gamma initpolicy using various shape and
scale parameters, for RNGSSE and MKL generators (Random123 samples gamma
as RNGSSE does, MKL has its own API for it).

This can now be used to initialize the van Slooten & Pope gamma
frequency PDF model for turbulent time scale, to be implemented as a
dissipation rate model.

User configuration example:

walker

  dissipation

    init jointgamma
    icgamma
      # mean = 1.0, variance = 0.25, setting shape = 4.0 and scale = 1.0
      # see https://en.wikipedia.org/wiki/Gamma_distribution
      gammapdf 4.0 0.25 end
    end

  end

end